+++ UserService/Controllers/ViewModels/Common/Quaternion.cs
@@ -2,10 +2,10 @@ namespace UserService.Controllers.ViewModels.Common
 {
     public class Quaternion
     {
-        public decimal X { get; set; }
-        public decimal Y { get; set; }
-        public decimal Z { get; set; }
-        public decimal W { get; set; }
+        public double X { get; set; }
+        public double Y { get; set; }
+        public double Z { get; set; }
+        public double W { get; set; }
 
         public Domains.ValueObject.Quaternion ToModel()
         {
