+++ UserService/Controllers/ViewModels/RequestModels/ReplaceCharacterRequestModel.cs
@@ -12,7 +12,7 @@ namespace UserService.Controllers.ViewModels.RequestModels
         public Position Position { get; set; }
         public Quaternion Quaternion { get; set; }
         public Attributes Attributes { get; set; }
-        public decimal Experience { get; set; }
+        public double Experience { get; set; }
     }
     public class ReplaceCharacterRequestModelValidator : AbstractValidator<ReplaceCharacterRequestModel>
     {
